requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-environ==0.9.0
requests==2.28.1
pymongo
mongoengine
djongo
dnspython
